feat(components): create new LabwareStackRender component (#15842)
# Overview

According to new modal designs for labware/adapter/module stacks on
desktop and ODD, we need to render an isometric view of each element of
the stack, maintaining the topography of each stacked element. The
isometric views are constructed by transforming and aligning the labware
SVG (top face) along with two new rects (front and left side). All three
are scaled, skewed, and rotated at 30˚ angles or its derivatives
([reference](http://jeroenhoek.nl/articles/svg-and-isometric-projection.html)).

The "bottom" labware definition is an optional prop, so we have an
affordance for rendering an isometric view of a single labware from the
same component. The user can choose to highlight both or none of the
labware in the stack, resulting in a fill of blue30.

In addition, adapters are handled in the following way:
- If the adapter is used as the top definition and there is no bottom
definition, we render the non-transformed `LabwareAdapter` svg
- if the adapter is used as the bottom definition, we render the
isometric transform of the well-less adapter face. _**NOTE**_ In this
PR, the adapter's height is maintained, but we may also consider setting
the height for a bottom adapter as a constant, rendering a
standard-height rectangular prism with well-less face.

Closes [PLAT-374](https://opentrons.atlassian.net/browse/PLAT-374),
[PLAT-377](https://opentrons.atlassian.net/browse/PLAT-377),
[PLAT-394](https://opentrons.atlassian.net/browse/PLAT-394),
[PLAT-375](https://opentrons.atlassian.net/browse/PLAT-375)