feat(app, components): add labware modal for stacked entities at labware setup step map view
To give clarity to the contents of labware/adapter/module stacks, here, I add a modal when clicking
a stack on Labware setup deck map (for both Desktop and ODD). Each element of the stack will be
highlighted described in a list item containing the element's name, optional nickname, and isometric
SVG or PNG representation depending on its type.

Closes PLAT-376, PLAT-378