feat(app): Display run setup task completion (#15889)
In the run page on both app and ODD, you get an indication of completion
when you get all your instruments and modules and deck stuff present and
calibrated. But those are just two of the steps presented equally in the
run setup page, and the rest of the steps don't get anything similar. It
leads people to wonder whether they've set things up properly.

This PR adds similar styling and completion semantics for the other
tasks in the run setup screen to fix this issue. Specifically, LPC gets
a "confirm offsets" button (which will confirm offsets even if you
haven't run LPC - makes it more apparent that that's a separate option)
and labware and liquids get generic confirm buttons.

There's also a couple other visual fixes:
- On desktop, the "back to top" button in run setup is now where figma
thinks it is, outside the run-setup content area. This allows some
refactoring of component props
- On desktop, there was an issue with the react-router upgrade (I think
- it's also in the latest IR alpha) that means that if you had an
ongoing run, you couldn't view anything but run details without getting
instantly navigated back to run details

This implements this figma:
https://www.figma.com/design/Rwdt9R0aERFC55oTLDTlqY/8.0-September-Release-File?node-id=39-35830&t=l6vwJjQsfyVeovfC-4

## To come out of draft
- [x] implement for ODD
- [x] rebase onto release
- [x] "are you sure" modal on desktop
- [x] "are you sure" modal on ODD

## Review requests
- This is some pretty complex UI - do you agree with how I've done this?
- Some of this is pretty ugly, in large part because this is old code
that I'm cleaning up. There's some duplicated logic in the run details
and some pretty ugly typing. What I'd like to do is merge this since it
implements some features nicely and then follow up with a refactor to
get the size of some of these files down and enforce nicer separation
between everything.

## Testing
- [x] Desktop green checks on flex
- [x] Desktop green checks on OT-2 (yes, this has to be different
because the steps can be different here)

Closes RSQ-7