refactor(api, robot-server): CSV parameter improvements (#15907)
# Overview

This PR follows up on #15855 with various small refactors/improvements.

The biggest change in this PR is refactoring how we open the CSV
parameter files. Previously we were opening them when we first
initialized the parameter. This proved to be a potential issue because
we were only closing the files when we reached the execute step of
running the protocol (i.e. in a `finally` clause within `run_protocol`),
yet there was the small yet possible chance of not reaching that step
(for instance if the initial home command failed unexpectedly). To solve
this, we are now passing the `Path` object to the `CSVParameter`
interface and only opening the file when it is first accessed in the
protocol, meaning it will _always_ occur within that try/except block.

Another change introduced is properly raising a 400 error for `POST`s to
the `/runs`, `/protocols` and `/protocols/{protocolId}/analyses`
endpoints when referencing data files (for runtime parameter files) that
don't exist.

API Version is now also enforced for using `add_csv_file` which now
requires your protocol to be at 2.20. The api version is also now passed
down to the `CSVParameter` interface, though it is not currently used to
gate anything yet.

## Test Plan and Hands on Testing

Tested end to end on hardware with the same protocol/csv files as
#15855.

## Changelog

- Pass `Path` to `CSVParameter` interface and open files within that.
- Raise a 400 error when referencing non existent data file IDs for
`/runs`, `/protocols` and `/protocols/{protocolId}/analyses`
- Require API version 2.20 for `add_csv_file` 
- Fixes/improvements for testing CSV files/file opening in protocol api

## Review requests

## Risk assessment

Low.