fix(app): fix liquid & modules confirm
Fixes RQA-2925 for desktop: the liquids entry now will count as complete
for the purposes of not popping or being listed in the modal if you have
no liquids. It also gets complete decorations if there are no
liquids. That required adding that rightElement thing to the empty
version of the setup step, and that also is what we would need to do for
modules, so also add decorations for modules and fix the completion
logic to count the module actually being plugged in again (this was
RQA-2928).

On the ODD, similar kind of logic but a lot easier to implement because
the data is more local; we can default the state to !hasLiquids, and
then use the same styling we already had implemented for runtime
parameters for the setup step.

Also, remove some stylistically out of date text transforms on the setup
cards.

Closes RQA-2928
Closes RQA-2925