fix(app): restore scroll to top when changing desktop routes
Seemingly after React Router migration from v5->v6, the window retains its scroll position even
after changing routes. We would expect scroll position to renew to the top of the window when
changing routes, as the `Route` component is updated. To force this, I pass a key to the
`React.Fragment` supplied as the `element` prop within each mapped `Route`, following the API. This
seems to be related to a separate top-level routing issue on the ODD and should be investigated
further after release.

RQA-2953