fix(api): make frontRightNozzle and backLeftNozzle required for Quadrant config (#16025)
Closes RQA-2992

# Overview

The `QuadrantNozzleLayoutConfiguration` has two optional fields-
`frontRightNozzle` and `backLeftNozzle`, but they didn't have a default
value assigned. So when the app tried to fetch the run's data by sending
a `GET /runs/{run_id}/commands/` request, if the `configureNozzleLayout`
command was saved as the following valid command in the database-

```
{
  "id": "44c2ebda-4e81-42f4-bbd7-91710ca30242",
  "createdAt": "2024-08-13T20:00:19.662654+00:00",
  "commandType": "configureNozzleLayout",
  "key": "2a57aa0e5f1e286b479c11d95e2cce35",
  "status": "succeeded",
  "params": {
    "pipetteId": "984f1433-acfd-48e7-b213-3c6fdfd85321",
    "configurationParams": {
      "style": "QUADRANT",
      "primaryNozzle": "H1",
      "backLeftNozzle": "B1"
    }
  },
  "result": {},
  "startedAt": "2024-08-13T20:00:19.664360+00:00",
  "completedAt": "2024-08-13T20:00:19.676488+00:00",
  "notes": []
}
```

..then pydantic ran into a validation error since it didn't know how to
parse `configurationParams` with a missing `frontRightNozzle`.

This was resulting in the 'Download Run Log' option to give a 500 error,
which this PR fixes by making the fields no longer optional.

## Changelog

- We could have made the default values `None` but that would require a
change in the command schema. So in order to not change the schema, and
also make the API a bit more strict while doing it, we decided to make
the fields non-optional.
This PR makes this change and updates the `configure_nozzle_layout` API
method so that a Quadrant configuration (used by the API-side
`PARTIAL_COLUMN` & `QUADRANT` configurations) will always set its
`front_right` & `back_left` nozzles.

- While making this change, I discovered a missing check on the `end`
param- that it should be only allowed for `PARTIAL_COLUMN`
configuration. While adding that I ended up refactoring the validation
check since it was bloating up even further.

- Similarly, adding more test cases was making the tests quite wordy and
taking up a lot of space in the 1k+ lines file. So I moved all the
partial tip test setup stuff to its own file and added more test cases

## Review requests

- Make sure I have accounted for all the existing validation errors on
`configure_nozzle_layout`
- I think this will not fix the issues with runs that were already run
and saved on the DB with the optional fields. But there should not be
any robots in the field that would have this problem because Quadrant
configuration was not available in the public API until now. So I think
this is okay but let me know if I am missing any other problem with this
approach.

## Risk assessment

Low. A bug fix and mostly refactoring of rest of the code.