docs(api): add deep well adapter for temperature module (#16037)
# Overview

List the new Opentrons 96 Deep Well Temperature Module Adapter in the
Python docs.

There are no `versionadded` statements or additions to the Versioning
page because adding labware _is not_ gated by API level.

Addresses RTC-471.

## Test Plan and Hands on Testing


[Sandbox](http://sandbox.docs.opentrons.com/docs-temp-mod-deep-well-adapter/v2/modules/temperature_module.html#standalone-adapters)

## Changelog

- add one row to table

## Review requests

- Anywhere else we would want to mention this?
- Do we need to say what labware is compatible with the adapter, and how
to make your own with stacking offsets? We don't do that for other
adapters but it seems to be a common source of confusion. Maybe better
addressed in a separate PR.

## Risk assessment

nada