fix(app): dismiss run on `AnalysisFailedModal` CTA
On ODD, because `TopLevelRedirects` navigates to the current run route if such a run exists, we need
to be sure to explicitly clear the current run when navigating away, or we will be immediately
bounced back to the current run screen. With the introduction of RTPs that will cause reanalysis to
fail at the run, we show a new `AnalysisFailedModal` component. The CTA button should first dismiss
the current run, then navigate to protocol details or protocols landing, ensuring this bounce does
not occur.

RQA-3004