fix(app): correct visuals for LPC errors (#16091)
For reasons I don't understand, the LPC fatal error screen is its own
modal, which maybe used to be good but now looks extremely weird on the
ODD and maybe the app. Maybe because we changed the modal background or
something? I don't know why we didn't see this before.

Anyway, it's gross that it works that way, so just split it into a
contents component and an actual modal component. When we get a fatal
error that's not bad enough to break the JS, we render it as the active
component, and when we hit the (unique?) error boundary, we render it as
a modal.

Closes RQA-2987