fix(robot-server): allow partial network info
We have a single route that gets all the network info for the machine.
If the system command to get info for one of the interfaces failed, the
route as a whole would fail. We don't want that, because we want to get
the rest of the information. Instead, let's catch teh error in the
general overview endpoint, and return whatever information we have.

Closes RQA-3000