fix(app): fix run context not closing on the OT-2 after cancelled run (#16112)
Closes RQA-3090 and RQA-3062

In #16069, we properly close the run context...for Flexes only. Because we don't do tip status checks on the OT-2 post-run, the initialPipettesWithTipsCount is never populated, so the run never closes. This leads to unexpected behavior, like not being able to calibrate your OT-2 anymore, do normal post-run things, etc.