fix(app): handle usb req errors in redux robotapi
More recent app code uses the api-client direct fetch interface through
react-api-client hooks to talk to the robot, but older app code like
that in the desktop wifi connection uses what came before, an rxjs based
api that you interact with through redux. This API doesn't seem to
properly handle the new errors from the USB interface, so fix that.

This happens in a pretty ugly way because the type interface doesn't
really have the concept of requests that fail without having an http
status code, so we fake one.

Closes RQA-3087