fix(app): remove module card navigate on settled run query
removes a navigate route change on a settled run query with no data. it's unclear what the original
purpose of the route change was, but the /upload route no longer exists and falls through to the
/protocols page. the onSettled callback is triggered inconsistently on some robots when an error
propagates, likely due to recent run query notification changes. resolving the cause of that error
should be considered in future notification refactor work

closes RQA-3099