fix(app): dont ignore errors in redux robotapi (#16128)
we added the ability for the USB interface to actually pass on
communication errors to the browser side, but only some of the things
that create those requests on the browser handled it properly. The old
redux robot api client didn't. Now it does, at least on USB.

This sort of fixes RQA-3087 in that it will no longer be in the spinner
forever and correctly expose the error that happened on the robot, but
it doesn't fix it in that you'll now get an error and then a moment
later the robot will say it's connected anyway.

Closes RQA-3087 (under the above constraint)