fix(app): fix users not directed to "run preview" when starting a run (#16116)
Closes RQA-3092

We want users to be directed to the run preview panel (from any tab, not just the setup tab) when a run has started. We still want to ensure that users can still navigate to all tabs (which is why the dependency array is the way it is). See the code comment above the useEffect for the intended behavior.