fix(app): filter robots for CSV protocols that raise analysis errors (#16145)
We are intentionally permissive of robots on which to start protocol
setup if analysis fails app-side. However, we previously checked whether
the analysis contained errors to make this determination. Given that CSV
parameter protocols raise analysis errors intentionally, but contain
useful information regarding the robot type, we should filter robots if
the analysis result is `parameter-value-required`.

Closes RQA-3114