fix(app): fix labware location on labware details modal (#16164)
On ProtocolSetupLabware, we open a labware modal, either single or
stacked, when a labware is clicked. That modal's header contains a
location icon that is meant to specify that labware's/stack's initial
position. The bug here arises because the location used for the modal
header is the analysis's labware location, which contains the labware's
final location. Here, I pull the location from load commands.

Closes RQA-2800, Closes RQA-3132