fix(app): command correct mounts in detach
When the pipette wizard flows - which we use both in the instruments
flows and in the protocol setup flows - are instantiated, they get a
mount, and then they were using that mount blindly for all commands.
That's not always correct, though - if there's a 96 channel attached,
and the pipette wizard flow is instantiated to make sure there's an 8
channel on the right mount, then it needs to move the left mount to the
maintenance position to detach the 96.

to fix this, add a util that can get the "covering pipette" - i.e., a 96
if there is one and whatever's on the mount otherwise - and use that for
the pre-detach commands rather than whatever might or might not be on
the target mount.

Closes RQA-3123