fix(app): handle detaching 96 when attaching right (#16174)
When we're switching pipettes from a preprotocol flow, we were not
properly handling the case where a 96 is attached and we want to take it
off and attach a pipette to the right mount. Add handling for that.

I'm not really sure how this survived so long because it sure seems like
it's always been there.

When reviewing, go commit by commit to see the flow of refactors. Also
the logic table looks awful in the diff view unless you make your screen
really wide but it's fine if you're just looking at the file

## testing
- [x] if you have a 96 channel attached, and you set up a protocol with
a multi or single on the right mount, does the flow for detaching the 96
and attaching the right mount pipette go correctly

Closes RQA-3123