feat(components, protocol-designer): create search bar for labware tools
Here, I create a new search bar for LabwareTools for filtering labware and conditionally rendering
accordion buttons based on whether or not they contain a match for your search filter. I refactor
the Toolbox component to accommodate an element as the title type so that we can pass it a deck
location icon, and fix buggy behavior with opening and collapsing labware group accordions.

Closes AUTH-710, Closes AUTH-749