fix(app): fix labware location on labware details modal (ODD protocol details)
On ODD Protocol Details > Deck tab, we render the protocol deck with clickable labware. Clicking a
labware or stack should render the appropriate modal, whose header should be the location of that
labware or stack. However, similar to how we handle this in Protocol Setup, we need to find the
__initial__ locaiton of the labware at load time, rather than the final location of the labware.
Here, I access that initial location to properly display the slot location icon for labware that is
moved.

Closes RQA-3149