fix(app): ApiHostProvider in protocols page run
So, we indicate that a robot is connected via USB by setting its ip to
'opentrons-usb'. However, this doesn't do anything on its own, beyond
making http requests to http://opentrons-usb:31950 fail in a funny way.
It only works if you're making a request inside a mounted
<ApiHostProvider> component, which will provide a react context provider
for all the useHost hooks mounted underneath it to go look up the right
request function to use.

This generally works on anything under the devices page. Where it most
certainly didn't work was anything under the Protocols page, including
the CHooseRobotToRunProtocolSlideout component, which is the menu you
get when you start a protocol from the protocols page.

This only wasn't _comprehensively_ broken because at various times
various people had noticed the problem for various HTTP requests and
added a specific host override to the mutations used there... but nobody
had gotten around to doing that for the queries that get old runs and
scrape LPC data for them, and therefore that would never happen.

Instead, let's put an ApiHostProvider with the appropriate parameters
over the component, rip out the per-query/per-mutation host overries,
and have them all work through the magic of nonlocal state.

Closes RQA-2723